refactor : 컴포넌트명.style.ts -> styles.ts로 파일명 변경
