+++ src/components/Sidebar/Weather/styles.ts
