+++ src/components/common/Box/index.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import { BoxStyled } from './Box.style';
+import { BoxStyled } from './styles';
 
 export const Box = ({ children = '' }: { children: ReactNode }) => {
   return <BoxStyled>{children}</BoxStyled>;
