+++ src/components/common/Button/index.ts
@@ -1,1 +1,1 @@
-export * from './Button.style';
+export * from './styles';
