+++ src/components/common/Card/styles.tsx
