+++ src/constants/index.ts
@@ -30,8 +30,6 @@ export const INITIAL_LOCATION = {
   y: 90,
 };
 
-export const CATEGORY_LIST = 'categoryList';
-
 export const MOCK_TODOS = [
   {
     category: 'Study',
